add zip extension; moves account_feedback; adds feedback tests
